bfix: handles error incase of unsuccessful logout (#2420)
* bfix: handles error incase of unsuccessful logout

* test: adds Test if Logout is successful test

* ptch: routes to login page after logout

* docs: changed a sentence to a better non reductant one

+++ lib/views/after_auth_screens/app_settings/app_settings_page.dart
@@ -302,11 +302,16 @@ class AppSettingsPage extends StatelessWidget {
                         successText: 'LogOut',
                         success: () async {
                           try {
-                            await model.logout();
+                            final bool isLogoutSuccessful =
+                                await model.logout();
+                            if (!isLogoutSuccessful) {
+                              throw Error(); //checks whether the logout was successful or not.
+                            }
                             navigationService.pop();
                             navigationService.removeAllAndPush(
-                              '/selectLang',
-                              '/',
+                              Routes.setUrlScreen,
+                              Routes.splashScreen,
+                              arguments: '',
                             );
                           } catch (e) {
                             navigationService.pushDialog(
